+++ diffractio/scalar_fields_X.py
@@ -81,7 +81,7 @@ from .__init__ import degrees, mm, np, plt
 
 from .config import Draw_X_Options, empty_types
 from .utils_typing import npt, Any, NDArray,  NDArrayFloat, NDArrayComplex
-from .utils_common import get_date, load_data_common, save_data_common, check_none, add
+from .utils_common import get_date, load_data_common, save_data_common, check_none, add, oversampling
 from .utils_drawing import normalize_draw
 from .utils_math import (fft_filter, get_edges, nearest, reduce_to_1, Bluestein_dft_x, get_k, nearest2)
 from .utils_multiprocessing import (_pickle_method, _unpickle_method,
@@ -244,6 +244,18 @@ class Scalar_field_X():
             return u_new
         else:
             self.u = np.conj(self.u)
+
+    @check_none('x','u', raise_exception=False)
+    def oversampling(self, factor_rate: int | tuple):
+        """Overfample function has been implemented in scalar X, XY, XZ, and XYZ frames reduce the pixel size of the masks and fields. 
+        This is also performed with the cut_resample function. However, this function oversamples with integer factors.
+        
+        Args:
+            factor_rate (int | tuple, optional): factor rate. Defaults to 2.
+        """
+
+        self = oversampling(self, factor_rate)
+
 
     def duplicate(self, clear: bool = False):
         """Duplicates the instance