+++ diffractio/scalar_fields_XZ.py
@@ -90,7 +90,7 @@ from .scalar_fields_X import (PWD_kernel, Scalar_field_X, WPM_schmidt_kernel,
                               kernelRS, kernelRSinverse)
 from .scalar_masks_X import Scalar_mask_X
 from .scalar_sources_X import Scalar_source_X
-from .utils_common import get_date, load_data_common, save_data_common, check_none
+from .utils_common import get_date, load_data_common, save_data_common, check_none, oversampling
 from .utils_drawing import normalize_draw, prepare_drawing
 from .utils_math import get_k, nearest, reduce_to_1, rotate_image
 from .utils_multiprocessing import _pickle_method, _unpickle_method
@@ -515,6 +515,18 @@ class Scalar_field_XZ():
                 self.__dict__ = dict0
             else:
                 raise Exception('no dictionary in load_data')
+
+
+    @check_none('x','z','u', raise_exception=False)
+    def oversampling(self, factor_rate: int | tuple):
+        """Overfample function has been implemented in scalar X, XY, XZ, and XYZ frames reduce the pixel size of the masks and fields. 
+        This is also performed with the cut_resample function. However, this function oversamples with integer factors.
+        
+        Args:
+            factor_rate (int | tuple, optional): factor rate. Defaults to 2.
+        """
+
+        self = oversampling(self, factor_rate)
 
     @check_none('x','z','u',raise_exception=False)
     def cut_resample(self,